requirements.txt
# MCP Server Dependencies
# Author: devsen

fastmcp>=0.1.0
sqlalchemy>=2.0.0
pyyaml>=6.0.0
requests>=2.31.0
fastapi>=0.95.0
uvicorn>=0.21.0
